Split RubyLexerBase features into separate traits
